api/requirements.txt
Django~=3.1.7
django-cors-headers~=3.7.0
djangorestframework~=3.12.2
djoser~=2.1.0
Pillow~=8.2.0
